ml-client/requirements.txt
boto3==1.26.32
filetype==1.2.0
Flask==2.2.2
numpy==1.20.1
Pillow==9.3.0
pymongo==4.3.3
pytest==6.2.3
python-dotenv==0.21.0
requests==2.25.1
tensorflow==2.11.0
tensorflow_hub==0.12.0
